Refactor lrowaccal app to be callable for testing and add radiance units to output PVL
