+++ isis/src/lro/apps/lrowaccal/lrowaccal.h
@@ -1,0 +1,33 @@
+#ifndef lrowaccal_h
+#define lrowaccal_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+#include <vector>
+#include <QString>
+#include <QDir>
+#include <QRegExp>
+
+#include "Brick.h"
+#include "Camera.h"
+#include "Constants.h"
+#include "CubeAttribute.h"
+#include "iTime.h"
+#include "Message.h"
+#include "NaifStatus.h"
+#include "Preference.h"
+#include "ProcessByBrick.h"
+#include "PvlGroup.h"
+#include "SpecialPixel.h"
+#include "Statistics.h"
+
+using namespace std;
+using namespace Isis;
+
+namespace Isis {
+  extern void lrowaccal(UserInterface &ui);
+  extern void lrowaccal(Cube *icube, UserInterface &ui, vector<QString> darkFiles, 
+    QString flatFile, QString radFile, QString specpixFile, QString tempFile);
+}
+
+#endif
